[ArticleBundle] Enhancement: Streamline tests
