+++ src/Kunstmaan/ArticleBundle/Tests/AdminList/AbstractArticlePageAdminListConfiguratorTest.php
