+++ src/Kunstmaan/ArticleBundle/Tests/DependencyInjection/KunstmaanArticleExtensionTest.php
@@ -6,9 +6,6 @@ use Kunstmaan\ArticleBundle\DependencyInjection\KunstmaanArticleExtension;
 use Matthias\SymfonyDependencyInjectionTest\PhpUnit\AbstractExtensionTestCase;
 use Symfony\Component\DependencyInjection\Extension\ExtensionInterface;
 
-/**
- * Class KunstmaanArticleExtensionTest
- */
 class KunstmaanArticleExtensionTest extends AbstractExtensionTestCase
 {
     /**
