+++ src/Kunstmaan/ArticleBundle/Tests/Entity/AbstractArticleOverviewPageTest.php
@@ -13,9 +13,6 @@ class ArticleOverViewPage extends AbstractArticleOverviewPage
     }
 }
 
-/**
- * Generated by PHPUnit_SkeletonGenerator on 2012-09-04 at 16:54:04.
- */
 class AbstractArticleOverviewPageTest extends TestCase
 {
     public function testGettersAndSetters()
