+++ src/Kunstmaan/ArticleBundle/Tests/Entity/AbstractArticlePageTest.php
@@ -11,9 +11,6 @@ class ArticlePage extends AbstractArticlePage
 {
 }
 
-/**
- * Generated by PHPUnit_SkeletonGenerator on 2012-09-04 at 16:54:04.
- */
 class AbstractArticlePageTest extends TestCase
 {
     public function testGettersAndSetters()
