+++ src/Kunstmaan/ArticleBundle/Tests/Entity/AbstractAuthorTest.php
@@ -5,9 +5,6 @@ namespace Kunstmaan\ArticleBundle\Tests\Entity;
 use Kunstmaan\ArticleBundle\Entity\AbstractAuthor;
 use PHPUnit\Framework\TestCase;
 
-/**
- * Generated by PHPUnit_SkeletonGenerator on 2012-09-04 at 16:54:04.
- */
 class AbstractAuthorTest extends TestCase
 {
     /**
