+++ src/Kunstmaan/ArticleBundle/Tests/Form/AbstractArticlePageAdminTypeTest.php
@@ -2,16 +2,12 @@
 
 namespace Kunstmaan\ArticleBundle\Tests\Form;
 
-use Kunstmaan\ArticleBundle\Form\AbstractArticleOverviewPageAdminType;
 use Kunstmaan\ArticleBundle\Form\AbstractArticlePageAdminType;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Form\FormBuilder;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
-/**
- * Class AbstractArticleOverviewPageAdminTypeTest
- */
-class AbstractArticleOverviewPageAdminTypeTest extends TestCase
+class AbstractArticlePageAdminTypeTest extends TestCase
 {
     /**
      * @var AbstractArticlePageAdminType
@@ -20,13 +16,13 @@ class AbstractArticleOverviewPageAdminTypeTest extends TestCase
 
     public function setUp()
     {
-        $entity = new AbstractArticleOverviewPageAdminType();
+        $entity = new AbstractArticlePageAdminType();
         $this->object = $entity;
     }
 
     public function testBlockPrefix()
     {
-        $this->assertEquals('AbstractArticleOverviewPage', $this->object->getBlockPrefix());
+        $this->assertEquals('AbstractArticlePage', $this->object->getBlockPrefix());
     }
 
     public function testConfigureOptions()
@@ -39,7 +35,7 @@ class AbstractArticleOverviewPageAdminTypeTest extends TestCase
     public function testBuildForm()
     {
         $builder = $this->createMock(FormBuilder::class);
-        $builder->expects($this->exactly(3))->method('add')->willReturn($builder);
+        $builder->expects($this->exactly(5))->method('add')->willReturn($builder);
 
         /* @var FormBuilder $builder */
         $this->object->buildForm($builder, []);
